+++ threedi_modelchecker/checks/factories.py
@@ -1,4 +1,3 @@
-from ..threedi_model import custom_types
 from .base import EnumCheck
 from .base import ForeignKeyCheck
 from .base import GeometryCheck
@@ -7,6 +6,7 @@ from .base import NotNullCheck
 from .base import TypeCheck
 from .base import UniqueCheck
 from geoalchemy2.types import Geometry
+from threedi_schema import custom_types
 
 
 def get_level(table, column, level_map):